requirements.txt
# requirements.txt
python-telegram-bot==20.7
flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0